remove unnecessary headers, fix buffer overflow in LinkedListTests

+++ C/LinkedList/src/LinkedListTestHelpers.h
@@ -2,7 +2,7 @@
 #define LINKEDLISTTESTHELPERS_H
 #include "LinkedList.h"
 
-void LinkedListTest_print(Head HEAD, char *result, unsigned long resultSize);
+void LinkedListTest_print(Head HEAD, char *result, int resultSize);
 
 void assert_LinkedList(Head HEAD, char *expectation);
